window-provider package create + refactor
package naming consistency
