+++ background/package.json
@@ -32,7 +32,8 @@
     "@ethersproject/transactions": "^5.4.0",
     "@ethersproject/web": "^5.4.0",
     "@tallyho/hd-keyring": "^0.3.2",
-    "@tallyho/tally-provider-bridge-shared": "0.0.1",
+    "@tallyho/provider-bridge-shared": "0.0.1",
+    "@tallyho/window-provider": "0.0.1",
     "@uniswap/token-lists": "^1.0.0-beta.25",
     "ajv": "^8.6.2",
     "ajv-formats": "^2.1.0",
