+++ package.json
@@ -37,14 +37,16 @@
     "ui",
     "background",
     "provider-bridge",
-    "provider-bridge-shared"
+    "provider-bridge-shared",
+    "window-provider"
   ],
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.2",
     "@tallyho/tally-background": "0.0.1",
     "@tallyho/tally-ui": "0.0.1",
-    "@tallyho/tally-provider-bridge": "0.0.1",
-    "@tallyho/tally-provider-bridge-shared": "0.0.1",
+    "@tallyho/provider-bridge": "0.0.1",
+    "@tallyho/provider-bridge-shared": "0.0.1",
+    "@tallyho/window-provider": "0.0.1",
     "buffer": "^6.0.3",
     "react": "^17.0.2",
     "webext-options-sync": "^2.0.1",
