+++ window-provider/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@tallyho/provider-bridge-shared",
+  "name": "@tallyho/window-provider",
   "version": "0.0.1",
-  "description": "Tally, the community owned and operated Web3 wallet: provider bridge to connect the in-page provider with the background script.",
+  "description": "Tally, the community owned and operated Web3 wallet: window provider is responsible for creating the in-page object for communication.",
   "main": "index.ts",
   "repository": "git@github.com:thesis/tally-extension.git",
   "author": "Greg Nagy <gergo.nagy@tally.cash>",
@@ -13,5 +13,8 @@
     "wallet",
     "web3",
     "dapp"
-  ]
+  ],
+  "dependencies": {
+    "@tallyho/provider-bridge-shared": "0.0.1"
+  }
 }
